% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_names.R
\name{gene_list_genome}
\alias{gene_list_genome}
\title{Resolve a genome name to the value gene-list.com expects}
\usage{
gene_list_genome(genome)
}
\arguments{
\item{genome}{A single genome name.}
}
\value{
The value to send to the site.
}
\description{
Accepts the common name, the scientific name, or the value the site itself
uses, in any case and with spaces or underscores: \code{"Rat"},
\code{"rattus norvegicus"} and \code{"rattus_norvegicus"} are the same
genome.
}
\examples{
gene_list_genome("Rat")
gene_list_genome("Homo sapiens")
gene_list_genome("bakers yeast")

}
